+++ airbyte-metrics/lib/src/main/java/io/airbyte/metrics/lib/NotImplementedMetricClient.java
@@ -5,13 +5,14 @@
 package io.airbyte.metrics.lib;
 
 /**
- * A mock implementation of MetricClient. Useful for users who do not have any metric client but
- * still want to use the functionality of airbyte.
+ * A mock implementation of MetricClient. Useful for users who do not have any metric client set up
+ * but still want to use the functionality of airbyte, or in a unit test where user calls the
+ * testing function but did not initialize the metric client in the first place.
  */
 public class NotImplementedMetricClient implements MetricClient {
 
   @Override
-  public void count(MetricsRegistry metric, double val, String... tags) {
+  public void count(MetricsRegistry metric, long val, String... tags) {
     // Not Implemented.
   }
